feat: kernel: NetworkKit improvements and signal system refactors.
