+++ dev/kernel/HALKit/ARM64/HalCoreInterruptHandler.cc
@@ -8,7 +8,7 @@
 #include <KernelKit/ProcessScheduler.h>
 #include <KernelKit/UserMgr.h>
 #include <NeKit/KString.h>
-#include <SignalKit/SignalGen.h>
+#include <SignalKit/Signals.h>
 
 EXTERN_C Kernel::Void int_handle_breakpoint(Kernel::UIntPtr rip);
 EXTERN_C BOOL         mp_handle_gic_interrupt_el0(Void);
